thymio/requirements.txt
cycler==0.10.0
dbus-python==1.2.8
graphviz==0.9
kiwisolver==1.0.1
matplotlib==3.0.0
neat-python==0.92
numpy==1.15.3
pycairo==1.17.1
PyGObject==3.30.1
pyparsing==2.2.2
python-dateutil==2.7.3
six==1.11.0
